Created NPCController.cs
Removed all movement/rotation methods from EnemyController.cs into NPCController.cs. This script can be the parent of any future Enemy/NPC classes.
Renamed EnemyController to Enemy
