+++ Assets/Scenes/zareth_spider_bot_scene.unity
@@ -365,6 +365,14 @@ PrefabInstance:
       value: 0
       objectReference: {fileID: 0}
     - target: {fileID: 7968484937769075984, guid: d6fc5d24129c83147867b38155e51c9b, type: 3}
+      propertyPath: maxIdleTime
+      value: 1.3
+      objectReference: {fileID: 0}
+    - target: {fileID: 7968484937769075984, guid: d6fc5d24129c83147867b38155e51c9b, type: 3}
+      propertyPath: minIdleTime
+      value: 0.5
+      objectReference: {fileID: 0}
+    - target: {fileID: 7968484937769075984, guid: d6fc5d24129c83147867b38155e51c9b, type: 3}
       propertyPath: startingPosition
       value: 2
       objectReference: {fileID: 0}
@@ -375,6 +383,14 @@ PrefabInstance:
     - target: {fileID: 7968484937769075984, guid: d6fc5d24129c83147867b38155e51c9b, type: 3}
       propertyPath: whatIsGround.m_Bits
       value: 128
+      objectReference: {fileID: 0}
+    - target: {fileID: 7968484937769075984, guid: d6fc5d24129c83147867b38155e51c9b, type: 3}
+      propertyPath: maxTimeBeforeNextIdle
+      value: 2.5
+      objectReference: {fileID: 0}
+    - target: {fileID: 7968484937769075984, guid: d6fc5d24129c83147867b38155e51c9b, type: 3}
+      propertyPath: minTimeBeforeNextIdle
+      value: 2
       objectReference: {fileID: 0}
     m_RemovedComponents: []
   m_SourcePrefab: {fileID: 100100000, guid: d6fc5d24129c83147867b38155e51c9b, type: 3}
@@ -1316,8 +1332,6 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: d63da978b6f61494680b63f10cd7359c, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  maxHp: 0
-  attackDamage: 0
   speed: 4
   gcRaycastDist: 1
   facingRight: 0
@@ -1327,10 +1341,12 @@ MonoBehaviour:
   whatIsWall:
     serializedVersion: 2
     m_Bits: 128
-  minTimeBeforeNextStop: 4
-  maxTimeBeforeNextStop: 4.3
-  minStopTime: 1.2
-  maxStopTime: 2.3
+  minTimeBeforeNextIdle: 4
+  maxTimeBeforeNextIdle: 5
+  minIdleTime: 1.5
+  maxIdleTime: 3
+  maxHp: 3
+  attackDamage: 1
 --- !u!50 &1861660045
 Rigidbody2D:
   serializedVersion: 4
