+++ Assets/Scripts/scr_NPCs.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 60af8af32baae994b871da8bbf626699
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
