+++ Assets/Scripts/scr_NPCs/NPCController.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: a561dec5686c7994cb0183b128a702b5
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
